better organized python code
